requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
google-genai
python-dotenv
